+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/matchers/InterfaceMatchers.java
@@ -1,4 +1,4 @@
-package org.batfish.datamodel;
+package org.batfish.datamodel.matchers;
 
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.equalTo;
@@ -7,66 +7,58 @@ import com.google.common.collect.ImmutableSet;
 import java.util.List;
 import java.util.Set;
 import javax.annotation.Nonnull;
-import org.hamcrest.FeatureMatcher;
+import org.batfish.datamodel.SourceNat;
+import org.batfish.datamodel.matchers.InterfaceMatchersImpl.HasDeclaredNames;
+import org.batfish.datamodel.matchers.InterfaceMatchersImpl.HasSourceNats;
+import org.batfish.datamodel.matchers.InterfaceMatchersImpl.IsActive;
 import org.hamcrest.Matcher;
 
 public final class InterfaceMatchers {
-  private static final class HasDeclaredNames extends FeatureMatcher<Interface, Set<String>> {
-    private HasDeclaredNames(@Nonnull Matcher<? super Set<String>> subMatcher) {
-      super(subMatcher, "declared names", "declared names");
-    }
 
-    @Override
-    protected Set<String> featureValueOf(Interface actual) {
-      return actual.getDeclaredNames();
-    }
-  }
-
-  private static final class HasSourceNats extends FeatureMatcher<Interface, List<SourceNat>> {
-    private HasSourceNats(@Nonnull Matcher<? super List<SourceNat>> subMatcher) {
-      super(subMatcher, "sourceNats", "sourceNats");
-    }
-
-    @Override
-    protected List<SourceNat> featureValueOf(Interface actual) {
-      return actual.getSourceNats();
-    }
-  }
-
-  private static final class IsActive extends FeatureMatcher<Interface, Boolean> {
-    private IsActive(@Nonnull Matcher<? super Boolean> subMatcher) {
-      super(subMatcher, "active", "active");
-    }
-
-    @Override
-    protected Boolean featureValueOf(Interface actual) {
-      return actual.getActive();
-    }
-  }
-
+  /**
+   * Provides a matcher that matches if the provided the interface's declared names comprise the set
+   * of unique strings in {@code expectedDeclaredNames}.
+   */
   public static HasDeclaredNames hasDeclaredNames(@Nonnull Iterable<String> expectedDeclaredNames) {
     return new HasDeclaredNames(
         containsInAnyOrder(ImmutableSet.copyOf(expectedDeclaredNames).toArray()));
   }
 
+  /**
+   * Provides a matcher that matches if the provided {@code subMatcher} matches the interface's
+   * declared names.
+   */
   public static HasDeclaredNames hasDeclaredNames(
       @Nonnull Matcher<? super Set<String>> subMatcher) {
     return new HasDeclaredNames(subMatcher);
   }
 
+  /**
+   * Provides a matcher that matches if the provided the interface's declared names comprise the set
+   * of unique strings in {@code expectedDeclaredNames}.
+   */
   public static HasDeclaredNames hasDeclaredNames(@Nonnull String... expectedDeclaredNames) {
     return new HasDeclaredNames(
         containsInAnyOrder(ImmutableSet.copyOf(expectedDeclaredNames).toArray()));
   }
 
+  /**
+   * Provides a matcher that matches if the provided {@code subMatcher} matches the interface's
+   * source NATs.
+   */
   public static HasSourceNats hasSourceNats(@Nonnull Matcher<? super List<SourceNat>> subMatcher) {
     return new HasSourceNats(subMatcher);
   }
 
+  /** Provides a matcher that matches if the interface is active. */
   public static IsActive isActive() {
     return new IsActive(equalTo(true));
   }
 
+  /**
+   * Provides a matcher that matches if the provided {@code subMatcher} matches the interface's
+   * active flag.
+   */
   public static IsActive isActive(@Nonnull Matcher<? super Boolean> subMatcher) {
     return new IsActive(subMatcher);
   }
