+++ projects/batfish/src/main/java/org/batfish/grammar/cisco/CiscoControlPlaneExtractor.java
@@ -1,6 +1,8 @@
 package org.batfish.grammar.cisco;
 
-import com.google.common.collect.ImmutableSet;
+import static java.util.Comparator.naturalOrder;
+
+import com.google.common.collect.ImmutableSortedSet;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.EnumSet;
@@ -1006,7 +1008,7 @@ public class CiscoControlPlaneExtractor extends CiscoParserBaseListener
       _w.pedantic("Interface: '" + name + "' altered more than once");
     }
     newInterface.setDeclaredNames(
-        ImmutableSet.<String>builder()
+        new ImmutableSortedSet.Builder<String>(naturalOrder())
             .addAll(newInterface.getDeclaredNames())
             .add(ctx.getText())
             .build());
